Moved TokenIcon to components
