+++ source/components/TokenIcon/index.js
